list notes by users

+++ backend/controllers/notes.js
@@ -1,20 +1,14 @@
 const notesRouter = require('express').Router()
 const Note = require('../models/note')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+notesRouter.use(userExtractor)
 
 notesRouter.get('/', async (request, response) => {
-  const notes = await Note
-    .find({}).populate('user', { username: 1, name: 1 })
+  const notes = await Note.find({ user: request.user._id }).populate('user', { username: 1, name: 1 })
   response.json(notes)
+
 })
 
 notesRouter.get('/:id', async (request, response) => {
@@ -30,27 +24,27 @@ notesRouter.get('/:id', async (request, response) => {
 notesRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if(!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-
-  const user = await User.findById(decodedToken.id)
-
   const note = new Note({
     content: body.content,
     important: body.important === undefined ? false : body.important,
-    user: user.id
+    user: request.user._id
   })
 
   const savedNote = await note.save()
-  user.notes = user.notes.concat(savedNote._id)
-  await user.save()
 
   response.status(201).json(savedNote)
 })
 
 notesRouter.delete('/:id', async (request, response) => {
+  const note = await Note.findById(request.params.id)
+  if (!note) {
+    return response.status(404).json({ error: 'Note not found' })
+  }
+
+  const user = await User.findById(note.user)
+  user.notes = user.notes.filter(n => n.toString() !== note.id.toString()) // remove note from user
+  await user.save()
+
   await Note.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
